requirements.txt
pkg-resources==0.0.0
PyGithub==1.29
PyYAML==3.12
